+++ include/gba/ppu.h
@@ -112,7 +112,7 @@ static_assert(sizeof(union oam_entry) == 3 * sizeof(uint16_t));
 
 struct ppu {
     /* The emulator's screen as it is being rendered. */
-    uint32_t framebuffer[GBA_SCREEN_WIDTH * GBA_SCREEN_HEIGHT];
+    uint16_t framebuffer[GBA_SCREEN_WIDTH * GBA_SCREEN_HEIGHT];
 
     /* Internal registers used for affine backgrounds */
     int32_t internal_px[2];
